#241 [B] Technical Documentation and Configuration
* Create technical documentation for setup and maintenance of DAS 2020 (within README.md)
* Update configuration files, such as .env to ensure we have all the necessary options (such as for postgres and mysql)
* Ensure that development such as run.sh and build_react_with_django.sh give warnings that they are purely development only scripts.
* Provide PDF version of technical documentation in README.pdf

+++ backend/run.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
+# DEVELOPMENT SCRIPT ONLY!
+
+
+printf "\033[1;33mTHIS SCRIPT IS FOR DEVELOPMENT ONLY\033[0m\n"
+printf "\033[1;33mAre you sure you would like to continue? (Y/n)\033[0m\n"
+read result
+if [[ ${result} != "Y" ]]; then
+  exit 0;
+fi
+
 echo "Deleting migrations directory"
 $(rm -r api/migrations) 2>&1 > /dev/null
